Moving slow tests to NuGet.Core.FuncTests
