+++ test/NuGet.Core.Tests/NuGet.Protocol.Tests/HttpSource/HttpRetryHandlerTests.cs
@@ -13,6 +13,7 @@ using System.Threading.Tasks;
 using NuGet.Common;
 using NuGet.Test.Server;
 using NuGet.Test.Utility;
+using Test.Utility;
 using Xunit;
 
 namespace NuGet.Protocol.Tests
@@ -23,79 +24,6 @@ namespace NuGet.Protocol.Tests
         private const string TestUrl = "https://test.local/test.json";
         private static readonly TimeSpan SmallTimeout = TimeSpan.FromMilliseconds(50);
         private static readonly TimeSpan LargeTimeout = TimeSpan.FromSeconds(5);
-
-        [Fact]
-        public async Task HttpRetryHandler_HandlesFailureToConnect()
-        {
-            // Arrange
-            var server = new NotListeningServer { Mode = TestServerMode.ConnectFailure };
-
-            // Act & Assert
-            var exception = await ThrowsException<HttpRequestException>(server);
-#if IS_CORECLR
-            Assert.NotNull(exception.InnerException);
-            if (!RuntimeEnvironmentHelper.IsWindows)
-            {
-                Assert.Equal("Couldn't connect to server", exception.InnerException.Message);
-            }
-            else
-            {
-                Assert.Equal("A connection with the server could not be established", exception.InnerException.Message);
-            }
-#else
-            var innerException = Assert.IsType<WebException>(exception.InnerException);
-            Assert.Equal(WebExceptionStatus.ConnectFailure, innerException.Status);
-#endif
-        }
-
-        [Fact]
-        public async Task HttpRetryHandler_HandlesInvalidProtocol()
-        {
-            // Arrange
-            var server = new TcpListenerServer { Mode = TestServerMode.ServerProtocolViolation };
-
-            // Act & Assert
-            var exception = await ThrowsException<HttpRequestException>(server);
-#if IS_CORECLR
-            if (!RuntimeEnvironmentHelper.IsWindows)
-            {
-                Assert.Null(exception.InnerException);
-                Assert.Equal("The server returned an invalid or unrecognized response.", exception.Message);
-            }
-            else
-            {
-                Assert.NotNull(exception.InnerException);
-                Assert.Equal("The server returned an invalid or unrecognized response", exception.InnerException.Message);
-            }
-#else
-            var innerException = Assert.IsType<WebException>(exception.InnerException);
-            Assert.Equal(WebExceptionStatus.ServerProtocolViolation, innerException.Status);
-#endif
-        }
-
-        [Fact]
-        public async Task HttpRetryHandler_HandlesNameResolutionFailure()
-        {
-            // Arrange
-            var server = new UnknownDnsServer { Mode = TestServerMode.NameResolutionFailure };
-
-            // Act & Assert
-            var exception = await ThrowsException<HttpRequestException>(server);
-#if IS_CORECLR
-            Assert.NotNull(exception.InnerException);
-            if (!RuntimeEnvironmentHelper.IsWindows)
-            {
-                Assert.Equal("Couldn't resolve host name", exception.InnerException.Message);
-            }
-            else
-            {
-                Assert.Equal("The server name or address could not be resolved", exception.InnerException.Message);
-            }
-#else
-            var innerException = Assert.IsType<WebException>(exception.InnerException);
-            Assert.Equal(WebExceptionStatus.NameResolutionFailure, innerException.Status);
-#endif
-        }
 
         [Fact]
         public async Task HttpRetryHandler_DifferentRequestInstanceEachTime()
@@ -109,7 +37,7 @@ namespace NuGet.Protocol.Tests
             };
 
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -129,51 +57,10 @@ namespace NuGet.Protocol.Tests
         }
 
         [Fact]
-        public async Task HttpRetryHandler_AppliesTimeoutToRequestsIndividually()
-        {
-            // Arrange
-
-            // 20 requests that take 250ms each for a total of 5 seconds (plus noise).
-            var requestDuration = TimeSpan.FromMilliseconds(250);
-            var maxTries = 20;
-
-            // Make the request timeout longer than each request duration but less than the total
-            // duration of all attempts.
-            var requestTimeout = TimeSpan.FromMilliseconds(4000);
-
-            int hits = 0;
-            Func<HttpRequestMessage, CancellationToken, Task<HttpResponseMessage>> handler = async (requestMessage, token) =>
-            {
-                hits++;
-                await Task.Delay(requestDuration);
-                return new HttpResponseMessage(HttpStatusCode.InternalServerError);
-            };
-
-            var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
-            var httpClient = new HttpClient(testHandler);
-            var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
-            {
-                MaxTries = maxTries,
-                RequestTimeout = requestTimeout,
-                RetryDelay = TimeSpan.Zero
-            };
-            var log = new TestLogger();
-
-            // Act
-            using (await retryHandler.SendAsync(request, log, CancellationToken.None))
-            {
-            }
-
-            // Assert
-            Assert.Equal(maxTries, hits);
-        }
-
-        [Fact]
         public async Task HttpRetryHandler_CancelsRequestAfterTimeout()
         {
             // Arrange
-            CancellationToken requestToken = CancellationToken.None;
+            var requestToken = CancellationToken.None;
             Func<HttpRequestMessage, CancellationToken, Task<HttpResponseMessage>> handler = async (requestMessage, token) =>
             {
                 requestToken = token;
@@ -182,7 +69,7 @@ namespace NuGet.Protocol.Tests
             };
 
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -214,7 +101,7 @@ namespace NuGet.Protocol.Tests
             };
 
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -245,7 +132,7 @@ namespace NuGet.Protocol.Tests
             var minTime = GetRetryMinTime(MaxTries, SmallTimeout);
 
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -285,7 +172,7 @@ namespace NuGet.Protocol.Tests
             };
             
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -317,7 +204,7 @@ namespace NuGet.Protocol.Tests
             };
 
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -349,7 +236,7 @@ namespace NuGet.Protocol.Tests
             };
 
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -388,7 +275,7 @@ namespace NuGet.Protocol.Tests
             };
 
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -437,7 +324,7 @@ namespace NuGet.Protocol.Tests
             };
 
             var retryHandler = new HttpRetryHandler();
-            var testHandler = new TestHandler(handler);
+            var testHandler = new HttpRetryTestHandler(handler);
             var httpClient = new HttpClient(testHandler);
             var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, TestUrl))
             {
@@ -457,70 +344,9 @@ namespace NuGet.Protocol.Tests
             }
         }
 
-        private static async Task<T> ThrowsException<T>(ITestServer server) where T : Exception
-        {
-            return await server.ExecuteAsync(async address =>
-            {
-                // Arrange
-                var retryHandler = new HttpRetryHandler();
-                var countingHandler = new CountingHandler { InnerHandler = new HttpClientHandler() };
-                var httpClient = new HttpClient(countingHandler);
-                var request = new HttpRetryHandlerRequest(httpClient, () => new HttpRequestMessage(HttpMethod.Get, address))
-                {
-                    MaxTries = 2,
-                    RetryDelay = TimeSpan.Zero
-                };
-
-                // Act
-                Func<Task> actionAsync = () => retryHandler.SendAsync(
-                    request,
-                    new TestLogger(),
-                    CancellationToken.None);
-
-                // Act & Assert
-                var exception = await Assert.ThrowsAsync<T>(actionAsync);
-                Assert.Equal(2, countingHandler.Hits);
-                return exception;
-            });
-        }
-
         private static TimeSpan GetRetryMinTime(int tries, TimeSpan retryDelay)
         {
             return TimeSpan.FromTicks((tries - 1) * retryDelay.Ticks);
-        }
-
-        private class CountingHandler : DelegatingHandler
-        {
-            public int Hits { get; private set; }
-
-            protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
-            {
-                Hits++;
-
-                return base.SendAsync(request, cancellationToken);
-            }
-        }
-
-        private class TestHandler : HttpMessageHandler
-        {
-            private readonly Func<HttpRequestMessage, CancellationToken, Task<HttpResponseMessage>> _handler;
-
-            public TestHandler(Func<HttpRequestMessage, HttpResponseMessage> handler)
-            {
-                _handler = (request, token) => Task.FromResult(handler(request));
-            }
-
-            public TestHandler(Func<HttpRequestMessage, CancellationToken, Task<HttpResponseMessage>> handler)
-            {
-                _handler = handler;
-            }
-
-            protected override async Task<HttpResponseMessage> SendAsync(
-                HttpRequestMessage request,
-                CancellationToken cancellationToken)
-            {
-                return await _handler(request, cancellationToken);
-            }
         }
     }
 }
